Day 17 (Part 1)
This solution only runs the first part in a reasonable amount of time (no
guarantee it even works on the second problem).

I've made many attempts so far, including a rewrite in Rust, but the reality is
the slowness is likely algorithmic in nature. There's some key bit that I've
missed to make this really fast.